--- a/cleanlab/count.py
+++ b/cleanlab/count.py
@@ -28,43 +28,40 @@ These are estimated from a classification dataset. This module considers two typ
 * multi-label classification where each example can be labeled as belonging to multiple classes (e.g. ``labels = [[1,2],[1],[0],[],...]``)
 """
 
-from sklearn.linear_model import LogisticRegression as LogReg
-from sklearn.model_selection import StratifiedKFold
-from sklearn.metrics import confusion_matrix
-import sklearn.base
-import numpy as np
 import warnings
-from typing import Tuple, Union, Optional
+from typing import Optional, Tuple, Union
 
-from cleanlab.typing import LabelLike
-from cleanlab.internal.multilabel_utils import stack_complement, get_onehot_num_classes
+import numpy as np
+import sklearn.base
+from sklearn.linear_model import LogisticRegression as LogReg
+from sklearn.metrics import confusion_matrix
+from sklearn.model_selection import StratifiedKFold
+
 from cleanlab.internal.constants import (
-    TINY_VALUE,
     CONFIDENT_THRESHOLDS_LOWER_BOUND,
     FLOATING_POINT_COMPARISON,
-)
-
-from cleanlab.internal.util import (
-    value_counts_fill_missing_classes,
-    clip_values,
-    clip_noise_rates,
-    round_preserving_row_totals,
-    append_extra_datapoint,
-    train_val_split,
-    get_num_classes,
-    get_unique_classes,
-    is_torch_dataset,
-    is_tensorflow_dataset,
+    TINY_VALUE,
 )
 from cleanlab.internal.latent_algebra import (
     compute_inv_noise_matrix,
-    compute_py,
     compute_noise_matrix_from_inverse,
+    compute_py,
 )
-from cleanlab.internal.validation import (
-    assert_valid_inputs,
-    labels_to_array,
+from cleanlab.internal.multilabel_utils import get_onehot_num_classes, stack_complement
+from cleanlab.internal.util import (
+    append_extra_datapoint,
+    clip_noise_rates,
+    clip_values,
+    get_num_classes,
+    get_unique_classes,
+    is_tensorflow_dataset,
+    is_torch_dataset,
+    round_preserving_row_totals,
+    train_val_split,
+    value_counts_fill_missing_classes,
 )
+from cleanlab.internal.validation import assert_valid_inputs, labels_to_array
+from cleanlab.typing import LabelLike
 
 
 def num_label_issues(
@@ -1458,7 +1455,7 @@ def get_confident_thresholds(
         #  this approach is that there will be no standard value returned for missing classes.
         labels = labels_to_array(labels)
         all_classes = range(pred_probs.shape[1])
-        unique_classes = get_unique_classes(labels)
+        unique_classes = get_unique_classes(labels, multi_label=multi_label)
         BIG_VALUE = 2
         confident_thresholds = [
             np.mean(pred_probs[:, k][labels == k]) if k in unique_classes else BIG_VALUE