--- a/cleanlab/segmentation/filter.py
+++ b/cleanlab/segmentation/filter.py
@@ -182,41 +182,53 @@ def find_label_issues(
     ranked_label_issues = lab.get_label_issues()
     ### End find_label_issues_batched() section
 
-    # Finding the right indicies
-    relative_index = ranked_label_issues % (h * w)
-    pixel_coor_i, pixel_coor_j = np.unravel_index(relative_index, (h, w))
-    image_number = ranked_label_issues // (h * w)
-
     # Upsample carefully maintaining indicies
     label_issues = np.full((num_image, h, w), False)
 
     # only want to call it an error if pred_probs doesnt match the label at those pixels
-    for i in range(0, image_number.shape[0], batch_size):
-        image_batch = image_number[i : i + batch_size]
-        batch_i = pixel_coor_i[i : i + batch_size]
-        batch_j = pixel_coor_j[i : i + batch_size]
-        label_issues[image_batch, batch_i, batch_j] = True
+    for i in range(0, ranked_label_issues.shape[0], batch_size):
+        issues_batch = ranked_label_issues[i : i + batch_size]
+        # Finding the right indicies
+        image_batch, batch_coor_i, batch_coor_j = _get_indexes_from_ranked_issues(
+            issues_batch, h, w
+        )
+        label_issues[image_batch, batch_coor_i, batch_coor_j] = True
         if downsample == 1:
             # check if pred_probs matches the label at those pixels
-            pred_argmax = np.argmax(pred_probs[image_batch, :, batch_i, batch_j], axis=1)
-            mask = pred_argmax == labels[image_batch, batch_i, batch_j]
-            label_issues[image_batch[mask], batch_i[mask], batch_j[mask]] = False
+            pred_argmax = np.argmax(pred_probs[image_batch, :, batch_coor_i, batch_coor_j], axis=1)
+            mask = pred_argmax == labels[image_batch, batch_coor_i, batch_coor_j]
+            label_issues[image_batch[mask], batch_coor_i[mask], batch_coor_j[mask]] = False
 
     if downsample != 1:
         label_issues = label_issues.repeat(downsample, axis=1).repeat(downsample, axis=2)
 
-        # Upsample the coordinates
-        upsampled_ii = pixel_coor_i * downsample
-        upsampled_jj = pixel_coor_j * downsample
-        # Iterate over the upsampled region
-        for i in range(downsample):
-            for j in range(downsample):
-                rows = upsampled_ii + i
-                cols = upsampled_jj + j
-                pred_argmax = np.argmax(pred_probs[image_number, :, rows, cols], axis=1)
-                # Check if the predicted class (argmax) at the identified issue location matches the true label
-                mask = pred_argmax == labels[image_number, rows, cols]
-                # If they match, set the corresponding entries in the label_issues array to False
-                label_issues[image_number[mask], rows[mask], cols[mask]] = False
+        for i in range(0, ranked_label_issues.shape[0], batch_size):
+            issues_batch = ranked_label_issues[i : i + batch_size]
+            image_batch, batch_coor_i, batch_coor_j = _get_indexes_from_ranked_issues(
+                issues_batch, h, w
+            )
+            # Upsample the coordinates
+            upsampled_ii = batch_coor_i * downsample
+            upsampled_jj = batch_coor_j * downsample
+            # Iterate over the upsampled region
+            for i in range(downsample):
+                for j in range(downsample):
+                    rows = upsampled_ii + i
+                    cols = upsampled_jj + j
+                    pred_argmax = np.argmax(pred_probs[image_batch, :, rows, cols], axis=1)
+                    # Check if the predicted class (argmax) at the identified issue location matches the true label
+                    mask = pred_argmax == labels[image_batch, rows, cols]
+                    # If they match, set the corresponding entries in the label_issues array to False
+                    label_issues[image_batch[mask], rows[mask], cols[mask]] = False
 
     return label_issues
+
+
+def _get_indexes_from_ranked_issues(
+    ranked_label_issues: np.ndarray, h: int, w: int
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    hw = h * w
+    relative_index = ranked_label_issues % hw
+    pixel_coor_i, pixel_coor_j = np.unravel_index(relative_index, (h, w))
+    image_batch = ranked_label_issues // hw
+    return image_batch, pixel_coor_i, pixel_coor_j